Refactored model to match DB

+++ Models/basic/Subscription.cs
@@ -9,7 +9,7 @@ public partial class Subscription
     public int subscription_id {get; set;}
     public string? subscription_name {get; set;}
 
-    public float? subscription_cost {get; set;}
+    public double? subscription_cost {get; set;}
 
     public required int quality_id {get; set;}
 }
